movel ogic ot comon class
