+++ app/src/main/kotlin/com/wire/android/ui/home/newconversation/AddMembersToConversationViewModel.kt
@@ -1,0 +1,80 @@
+package com.wire.android.ui.home.newconversation
+
+import androidx.lifecycle.SavedStateHandle
+import com.wire.android.R
+import com.wire.android.mapper.ContactMapper
+import com.wire.android.navigation.EXTRA_CONVERSATION_ID
+import com.wire.android.navigation.NavigationManager
+import com.wire.android.ui.home.newconversation.search.SearchResultState
+import com.wire.android.util.dispatchers.DispatcherProvider
+import com.wire.kalium.logic.data.id.QualifiedID
+import com.wire.kalium.logic.data.id.parseIntoQualifiedID
+import com.wire.kalium.logic.data.user.UserId
+import com.wire.kalium.logic.feature.connection.SendConnectionRequestUseCase
+import com.wire.kalium.logic.feature.conversation.AddMemberToConversationUseCase
+import com.wire.kalium.logic.feature.conversation.GetAllContactsNotInConversationUseCase
+import com.wire.kalium.logic.feature.publicuser.Result
+import com.wire.kalium.logic.feature.publicuser.SearchKnownUsersUseCase
+import com.wire.kalium.logic.feature.publicuser.SearchUsersUseCase
+import javax.inject.Inject
+import com.wire.kalium.logic.feature.conversation.Result as GetContactsResult
+
+class AddMembersToConversationViewModel @Inject constructor(
+    private val getAllContactsNotInConversation: GetAllContactsNotInConversationUseCase,
+    private val searchKnownUsers: SearchKnownUsersUseCase,
+    private val searchPublicUsers: SearchUsersUseCase,
+    private val contactMapper: ContactMapper,
+    private val addMemberToConversation: AddMemberToConversationUseCase,
+    sendConnectionRequest: SendConnectionRequestUseCase,
+    savedStateHandle: SavedStateHandle,
+    navigationManager: NavigationManager,
+    dispatchers: DispatcherProvider
+) : SearchConversationViewModel(navigationManager, sendConnectionRequest, dispatchers) {
+
+    val conversationId: QualifiedID = savedStateHandle
+        .get<String>(EXTRA_CONVERSATION_ID)!!
+        .parseIntoQualifiedID()
+
+    override suspend fun getAllUsersUseCase() =
+        when (val result = getAllContactsNotInConversation(conversationId)) {
+            is GetContactsResult.Failure -> SearchResultState.Failure(R.string.label_general_error)
+            is GetContactsResult.Success -> SearchResultState.Success(
+                result.contactsNotInConversation.map { otherUser ->
+                    contactMapper.fromOtherUser(
+                        otherUser
+                    )
+                }
+            )
+        }
+
+    override suspend fun searchKnownUsersUseCase(searchTerm: String) =
+        when (val result = searchKnownUsers(searchTerm)) {
+            is Result.Failure.Generic, Result.Failure.InvalidRequest -> {
+                SearchResultState.Failure(R.string.label_general_error)
+            }
+            is Result.Failure.InvalidQuery -> {
+                SearchResultState.Failure(R.string.label_no_results_found)
+            }
+            is Result.Success -> {
+                SearchResultState.Success(result.userSearchResult.result.map { otherUser -> contactMapper.fromOtherUser(otherUser) })
+            }
+        }
+
+    override suspend fun searchPublicUsersUseCase(searchTerm: String): SearchResultState =
+        when (val result = searchPublicUsers(searchTerm)) {
+            is Result.Failure.Generic, Result.Failure.InvalidRequest -> {
+                SearchResultState.Failure(R.string.label_general_error)
+            }
+            is Result.Failure.InvalidQuery -> {
+                SearchResultState.Failure(R.string.label_no_results_found)
+            }
+            is Result.Success -> {
+                SearchResultState.Success(result.userSearchResult.result.map { otherUser -> contactMapper.fromOtherUser(otherUser) })
+            }
+        }
+
+    suspend fun addSelectedMembers() {
+        addMemberToConversation(conversationId, state.contactsAddedToGroup.map { UserId(it.id, it.domain) })
+    }
+
+}
